+++ tsdoc/src/nodes/DocParagraph.ts
@@ -1,10 +1,10 @@
-import { DocNodeKind } from './DocNode';
-import { DocSection, IDocSectionParameters } from './DocSection';
+import { DocNodeKind, DocNode } from './DocNode';
+import { DocNodeContainer, IDocNodeContainerParameters } from './DocNodeContainer';
 
 /**
  * Constructor parameters for {@link DocParagraph}.
  */
-export interface IDocParagraphParameters extends IDocSectionParameters {
+export interface IDocParagraphParameters extends IDocNodeContainerParameters {
 }
 
 /**
@@ -12,7 +12,7 @@ export interface IDocParagraphParameters extends IDocSectionParameters {
  * Like CommonMark, the TSDoc syntax uses blank lines to delineate paragraphs
  * instead of explicitly notating them.
  */
-export class DocParagraph extends DocSection {
+export class DocParagraph extends DocNodeContainer {
   /** {@inheritdoc} */
   public readonly kind: DocNodeKind = DocNodeKind.Paragraph;
 
@@ -22,5 +22,26 @@ export class DocParagraph extends DocSection {
    */
   public constructor(parameters: IDocParagraphParameters) {
     super(parameters);
+  }
+
+  /**
+   * {@inheritdoc}
+   * @override
+   */
+  public isAllowedChildNode(docNode: DocNode): boolean {
+    // NOTE: DocNodeKind.Paragraph cannot be nested
+    switch (docNode.kind) {
+      case DocNodeKind.BlockTag:
+      case DocNodeKind.CodeSpan:
+      case DocNodeKind.ErrorText:
+      case DocNodeKind.EscapedText:
+      case DocNodeKind.HtmlStartTag:
+      case DocNodeKind.HtmlEndTag:
+      case DocNodeKind.InlineTag:
+      case DocNodeKind.PlainText:
+      case DocNodeKind.SoftBreak:
+        return true;
+    }
+    return false;
   }
 }